requirements.txt
youtube-dl @ git+https://github.com/ytdl-org/youtube-dl.git@66ab0814c4baa2dc79c2dd5287bc0ad61a37c5b9
opencv-python
tensorflow-cpu
tensorflow_hub
matplotlib
pandas